requirements.txt
Flask
Pillow
ultralytics
torch
torchvision
torchaudio
python-dotenv